GCC warning fix

+++ src/server/game/Movement/MovementGenerators/PointMovementGenerator.cpp
@@ -196,7 +196,7 @@ void PointMovementGenerator::Finalize(Unit* owner, bool active, bool movementInf
 void PointMovementGenerator::MovementInform(Unit* owner)
 {
     // deliver EVENT_CHARGE to scripts, EVENT_CHARGE_PREPATH is just internal implementation detail of this movement generator
-    uint32 movementId = _movementId == EVENT_CHARGE_PREPATH ? EVENT_CHARGE : _movementId;
+    uint32 movementId = _movementId == EVENT_CHARGE_PREPATH ? uint32(EVENT_CHARGE) : _movementId;
 
     if (Creature* creature = owner->ToCreature())
         if (creature->AI())
